requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
diskcache==5.6.3
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.15.0
numpy==2.3.5
packaging==24.2
pandas==2.3.3
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==5.29.5
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
streamlit==1.37.1
streamlit-sortables==0.2.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
wheel==0.45.1

